feat(views): move topic flow endpoint to topic_flow view module
