+++ litigant_portal/app/views/pages.py
@@ -20,15 +20,11 @@ from litigant_portal.app.models.choices import (
     State,
     get_default_model,
 )
-from litigant_portal.app.selectors.admin import site_get_active_topics
-from litigant_portal.app.services.admin import (
+from litigant_portal.app.selectors.site import site_get_active_topics
+from litigant_portal.app.services.user import (
     user_can_access_admin,
 )
 from litigant_portal.app.topic_flow.answer_store import AnswerStore
-from litigant_portal.app.topic_flow.downloads import (
-    build_download,
-    find_downloadable,
-)
 from litigant_portal.app.topic_flow.registry import registry
 from litigant_portal.app.topic_flow.renderer import (
     question_ids,
@@ -152,33 +148,6 @@ def _render_topic_flow(request, corpus, answers, errors=None):
             "toc": toc,
         },
     )
-
-
-def topic_flow_download(request, court, topic, role, output_id):
-    """Download a Topic Flow output section as a file (e.g. an ``.ics``).
-
-    The generic counterpart to ``topic_flow``: resolve the corpus and the
-    downloadable output section (404 on either miss — an unknown id or a
-    non-downloadable section), then dispatch on ``output_type`` to assemble the
-    file from the guest's stored answers. The view stays thin — file bytes come
-    from the download handlers in downloads.py, computed from the same
-    AnswerStore the page renders, so the download matches what's on screen.
-    """
-    corpus = registry.get(court, topic, role)
-    if corpus is None:
-        raise Http404(f"No Topic Flow for {court}/{topic}/{role}")
-
-    section = find_downloadable(corpus, output_id)
-    if section is None:
-        raise Http404(f"No downloadable output {output_id!r}")
-
-    store = AnswerStore(request.session, court, topic, role)
-    artifact = build_download(section, corpus, store.all())
-    response = HttpResponse(artifact.body, content_type=artifact.content_type)
-    response["Content-Disposition"] = (
-        f'attachment; filename="{artifact.filename}"'
-    )
-    return response
 
 
 def about(request):
